Streamline GitLab API URL variables
The Platform and the Janitor now share a single variable in the Helm values file `global.gitlabApiUrl` for the GitLab URL. The handling of the GitLab API secret has been reworked too. The Secret resource has been renamed from nmaas-gitlab-janitor-token to nmaas-gitlab-api-token. Additionally it has been moved to `global.gitlabApiToken`. Old deployment can choose to manually specify `global.gitlabApiToken.secret.name` to the old value of `nmaas-gitlab-janitor-token` in order not to create a new secret.